Refactored code to move JAVA components to application_identifiers_java.py file
